nit typing
