+++ web/src/lib/types.ts
@@ -271,16 +271,21 @@ export interface SlackChannelDescriptor {
   name: string;
 }
 
-export interface SlackBot {
+export type SlackBot = {
   id: number;
   name: string;
   enabled: boolean;
   configs_count: number;
-
-  // tokens
+  slack_channel_configs: Array<{
+    id: number;
+    is_default: boolean;
+    channel_config: {
+      channel_name: string;
+    };
+  }>;
   bot_token: string;
   app_token: string;
-}
+};
 
 export interface SlackBotTokens {
   bot_token: string;
